improve spacing of top logo, topics, keywords
These changes do a decent job of moving the logo to align with the top of the printing on additional pages and separate it from the article title.  If you prefer that the logo stay where it was, and the article title move down -- that is easy enough (please let me know).  The spacing between the Abstract and the keywords or the topics and keywords is more natural.  There is one __flaw__:  if the author specifies topics but does not specify any keywords, then there is a `!` that appears after the topics ... I do not know what causes that.
Again .. let me know if you prefer the prior spacing if it does not make the `!` appear [I did not check for that].