move logo up, improve spacing for topics, keywords
Try it, see if you prefer this.  If you want the logo to stay where it was and just move the title down, let me know. 
There is one issue:  if the author specifies topics and does not specify keywords, then a '!' appears after the topics -- I do not know the cause or the fix.